Refactor CDLibrarian
