+++ Frontend/src/components/common/Pagination.tsx
@@ -3,7 +3,6 @@ import { View, TouchableOpacity, Text, StyleSheet } from "react-native";
 
 interface PaginationProps {
   currentPage: number;
-  isAtLastPage: boolean;
   goToPreviousPage: () => void;
   goToNextPage: () => void;
   lastPage: number;
@@ -11,7 +10,6 @@ interface PaginationProps {
 
 const Pagination: React.FC<PaginationProps> = ({
   currentPage,
-  isAtLastPage,
   goToPreviousPage,
   goToNextPage,
   lastPage,
