+++ Frontend/src/hooks/usePagination.tsx
@@ -1,9 +1,8 @@
 import { useState } from "react";
 
 const usePagination = () => {
-  const [isAtLastPage, setIsAtLastPage] = useState(false);
   const [currentPage, setCurrentPage] = useState(0);
-  const [totalPages, setTotalPages] = useState(0);
+  const [totalPages] = useState(0);
 
   const goToPreviousPage = () => {
     if (currentPage > 0) {
@@ -12,19 +11,15 @@ const usePagination = () => {
   };
 
   const goToNextPage = () => {
-    if (!isAtLastPage && totalPages == currentPage) {
+    if (totalPages == currentPage) {
       setCurrentPage(currentPage + 1);
     }
   };
   return {
     goToPreviousPage,
     goToNextPage,
-    setIsAtLastPage,
     setCurrentPage,
     currentPage,
-    isAtLastPage,
-    totalPages,
-    setTotalPages,
   };
 };
