+++ Frontend/src/services/pointsExchangeService.ts
@@ -22,5 +22,4 @@ export class PointsExchangeService {
       token: accessToken,
     });
   }
-
 }
